Move model extensions to pink.model
